pyproject.toml
[tool.poetry]
name = "blog-writer"
version = "0.1.0"
description = "Blog writer"
authors = ["thinhda"]
readme = "README.md"
packages = [{include = "blog_writer"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.263"
python-dotenv = "^1.0.0"
openai = "^0.27.8"
duckduckgo-search = "^3.8.4"
bs4 = "^0.0.1"
streamlit = "^1.25.0"
google-api-python-client = "^2.96.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
